Login: use settings.json
Let's add an untracked file called `settings.json` to store the username
and password that we will use to login.

Since we need the file for the user's username and password, let's print
an error message and exit if the file doesn't exist.

To login, we select the username and password fields by id and fill
them. Then we find the submit button within the login form and submit
it.