Navigation: go to Make a Tee Time
After logging in, a login confirmation appears for a couple seconds
until the user is redirected. To go to the page we want, we first need
to wait for this redirection. For no particular reason, let's wait for
the `.topnav_item` element to be present before going to the page we
need.