Navigation: open particular day's tee sheet
Let's add three more configurations in our `settings.json` file. A day
and month to book the tee time for and a earliest time to book for that
day.

To get to the particular day's tee sheet we're looking for, we need to
first find the month date picker for the month we specified. If that
month isn't found on the page, we print an error and exit.

Similarly, once we find a month we look for the day we specified. If
that day isn't a selectable day, we print an error and exit. If it is,
we click it to open that day's tee sheet.