Tee time registration: select players
We fist need to wait for the players to appear in the user's "Name List"
since this is poulated shortly after page load. Once ready, we can find
the options elements that correspond to the players specified in
settings. If there is a player that is specified in settings but does
not appear in the user's name list, an error message is printed and that
player is ignored.